+++ src/lesson.ts
@@ -30,6 +30,7 @@ b_replay.addEventListener("click",e=>{
         lesson.currentSubTask.replay(lesson.tut.getCurEditor()); // this needs to be fixed later; need to store on task _preTutFile and _prePFile to switch back to
         // ^^^ this has now been fixed but needs more fixing, need to store temporary state each time? uses a lot of ram but hmm
         // ^^^ since the core change the system has been remastered but still may need the temporary state system
+        // ^^^ it has been done
     }
 });
 b_goBackStep.addEventListener("click",e=>{